finish v0.3.1+2 version. Improve the download process and repair the problem that incremental update files do not exist.

+++ android/src/main/java/com/example/r_upgrade/common/IncrementUpgradeManager.java
@@ -25,12 +25,19 @@ public class IncrementUpgradeManager extends ContextWrapper {
         File parentFile = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS);
         File newApkFile = new File(parentFile, oldApkPath.substring(oldApkPath.lastIndexOf("/") + 1));
         try {
+            File patchFile = new File(patchPath);
+            if (!patchFile.exists()) {
+                return null;
+            }
+            if (patchFile.length() == 0) {
+                return null;
+            }
             if (newApkFile.exists()) {
                 newApkFile.delete();
             }
             newApkFile.createNewFile();
             rUpgradeLib.mixinPatch(oldApkPath, patchPath, newApkFile.getPath());
-            RUpgradeLogger.get().d(TAG, "mixinAndGetNewApk"+newApkFile.getPath());
+            RUpgradeLogger.get().d(TAG, "mixinAndGetNewApk" + newApkFile.getPath());
             return newApkFile.getPath();
         } catch (Exception e) {
             RUpgradeLogger.get().d(TAG, "合成失败：");
